requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.33.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.13.1
pydantic[email]==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.93.0
python-dateutil==2.9.0.post0
PyJWT==2.8.0
cryptography==41.0.7
Jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
